Arreglo Ejercicio 16 y 17

+++ practicas/Entrega1/Ejercicio16/Ejercicio16.js
@@ -3,8 +3,8 @@ function sumar(){
     let operando2 = document.getElementById('op2').value
     let result = "Debe ingresar 2 numeros"; 
     if ((!isNaN(operando1)) && (!isNaN(operando2)) && ((operando1 != "") && (operando2 != ""))){ //Me aseguro que sea un numero y que no este vacio
-        operando1 = parseInt (operando1,10);
-        operando2 = parseInt (operando2,10);
+        operando1 = parseFloat(operando1,10);
+        operando2 = parseFloat(operando2,10);
         result = operando1 + operando2;
     }
     document.getElementById('result').value = result;
