+++ practicas/Entrega1/Ejercicio17/Ejercicio17.js
@@ -8,8 +8,8 @@ function analizar(id){
     let result = "Debe ingresar 2 números";   
 
     if ((!isNaN(operando1)) && (!isNaN(operando2)) && ((operando1 != "") && (operando2 != ""))){ //Me aseguro que sea un numero y que no este vacio
-        operando1 = parseInt (operando1,10);
-        operando2 = parseInt (operando2,10);;
+        operando1 = parseFloat(operando1,10);
+        operando2 = parseFloat(operando2,10);;
 
         switch (id) {
             case "mulButton":
